fix for Issue-3
The fix i propose is simply to pass the epoch from the initial unstake tx to the callback.
Despite the suggestion to "Update pool.lastUnstake before the cross contract invocation.", I think it's better to update it in the callback, to ensure that no update is made if the unstake fails.
And because we are sequencing all unstake transactions, there can only be one of these that traverses multiple epochs.