fix for Issue-5
This fix ensures that the share price does not increases between the initial unstake request (where user shares are burned) and the unstake callback (where total staked is updated).

We're also adding some extra checks to the allocate, distribute_rewards, distribute_all functions to make sure they are executed only when the share price is up to date, and no stake/unstake transactions are in flight.

We don't think we need to lock these functions because they don't have callbacks, and they don't update the share price.