(non-)fix for Issue-10
We have not been able to reproduce this issue in a test, so we're just adding a test where the distributor provides the exact distribution amounts calculated by get_rewards_distribution_amounts, and successfully distributes in NEAR to many recipients.

We tried with different numbers of recipients, from 1 to 12 and the distribution always succeeds.
This test takes proportionally long to run, depending on how many recipients we add, so we're just leaving the test with 5 recipients.

The test confirms that the calculated TruNEAR amount can be slightly overestimated, and in fact with 5 recipients Alice is left with 2 yoctoNEAR of shares in excess to what it was calculated she would pay in distribution fees.

On the other hand, the amount of calculated NEAR is not underestimated, and the the distribution ends up spending no more that the value calculated.
For larger number of recipients the actual amount of NEAR spent can actually be a few yoctoNEAR less than it was calculated.

We believe this is because the NEAR amount calculated inside internal_distribute, that will be transferred to a recipient, gets rounded down multiple times (one per recipient) during the distribute_all loop.

```
// calculate the amount of rewards in NEAR
        let near_amount = NearToken::from_yoctonear(Self::convert_to_assets(
            shares_to_move,
            global_price_num,
            global_price_denom,
            false,
        ));
```