fix for Recommendation-5
Adding a test to show that when distribute_all in NEAR fails because of insufficient gas, the changes to the share price of the allocations processed so far are rolled back.
We already had a similar test to show that in the same situation of insufficient gas no event was emitted.

This behaviour makes sense because the updates to allocations share price are local to this contract, and all happen in the same transaction. Therefore they should be reverted if the transaction fails.